+++ __tests__/unit/core-transactions/handler-registry.test.ts
