+++ packages/core-transactions/src/index.ts
@@ -4,4 +4,4 @@ export * from "./services/transaction";
 export * from "./interfaces";
 
 export { errors };
-export { transactionHandlerRegistry as TransactionHandlerRegistry } from "./service-registry";
+export { transactionHandlerRegistry as TransactionHandlerRegistry } from "./handler-registry";
